refactor: more ui layer updates
