+++ app/src/main/AndroidManifest.xml
@@ -73,14 +73,14 @@
         <activity android:name=".ui.features.groups.members.FriendsAddActivity" />
         <activity android:name=".ui.features.groups.managment.CreateGroupActivity" />
         <activity android:name=".ui.features.groups.managment.AdminOptionsActivity" />
-        <activity android:name=".ui.features.groups.ChangeDateActivity" />
+        <activity android:name=".ui.features.groups.managment.ChangeDateActivity" />
         <activity android:name=".ui.features.groups.chat.ChatActivity" />
         <activity android:name=".ui.features.groups.members.MembersComingActivity" />
         <activity android:name=".ui.features.groups.members.MembersInvitedActivity" />
         <activity
             android:name=".ui.features.groups.members.FriendsRemoveActivity"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".ui.features.groups.UsersListActivity" />
+        <activity android:name=".ui.features.groups.members.UsersListActivity" />
         <activity android:name=".ui.features.groups.discovery.JoinGroupActivity" />
         <activity android:name=".ui.features.groups.managment.AdminSettingsActivity" />
         <activity android:name=".ui.features.groups.main.PartyMainActivity" />
