+++ app/src/main/java/com/example/partymaker/ui/features/groups/members/FriendsAddActivity.java
@@ -16,7 +16,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.example.partymaker.R;
 import com.example.partymaker.data.api.FirebaseServerClient;
 import com.example.partymaker.data.model.User;
-import com.example.partymaker.ui.features.groups.UsersListActivity;
 import com.example.partymaker.ui.features.groups.main.PartyMainActivity;
 import com.example.partymaker.utils.core.IntentExtrasManager;
 import com.example.partymaker.utils.core.ExtrasMetadata;
