+++ app/src/main/res/layout/list_party_friends.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context="com.example.partymaker.ui.features.groups.UsersListActivity">
+    tools:context="com.example.partymaker.ui.features.groups.members.UsersListActivity">
 
     <ListView
         android:id="@+id/lv"
